mjs_add_baseline {metricsgraphics}	R Documentation

Sets a baseline line/label

Description

metricsgraphics baselines are horizontal lines that may specify, say, a goal or target to be reached. This function lets you add baselines to a plot object. you can add as many as you need to.

Usage

mjs_add_baseline(mjs, y_value, label)

Arguments

mjs
metricsgraphics plot object

y_value
which y value to draw the baseline at

label
text label for the marker

Value

metricsgraphics object

Examples

data.frame(
  year=seq(1790, 1970, 10),
  uspop=as.numeric(uspop)
) %>%
  mjs_plot(x=year, y=uspop) %>%
  mjs_line() %>%
  mjs_add_marker(1850, "Something Wonderful") %>%
  mjs_add_baseline(150, "Something Awful")

[Package metricsgraphics version 0.9.1 Index]