mjs_add_legend {metricsgraphics}	R Documentation

Adds a legend to a metricsgraphics chart

Description

Adds a legend to a metricsgraphics chart

Usage

mjs_add_legend(mjs, legend, inline = FALSE)

Arguments

mjs
metricsgraphics plot object

legend
character vector of labels for the legend

inline
TRUE if you want line labes to the right of the chart vs in a legend block (experimental)

Value

metricsgraphics object

Examples

set.seed(1492)
stocks <- data.frame(
  time = as.Date('2009-01-01') + 0:9,
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4))

stocks %>%
  mjs_plot(x=time, y=X) %>%
  mjs_line() %>%
  mjs_add_line(Y) %>%
  mjs_add_line(Z) %>%
  mjs_axis_x(xax_format="date") %>%
  mjs_add_legend(legend=c("X", "Y", "Z"))

[Package metricsgraphics version 0.9.1 Index]