mjs_axis_x {metricsgraphics}	R Documentation

Configure x axis ticks & limits

Description

Configure x axis ticks & limits

Usage

mjs_axis_x(mjs, show = TRUE, xax_count = 6, min_x = NULL, max_x = NULL,
  extended_ticks = FALSE, xax_format = "plain",
  show_secondary_x_label = NULL, rug = FALSE)

Arguments

mjs
metricsgraphics plot object

show
display the axis? (default: TRUE - yes)

xax_count
tick count

min_x
min limit for x axis

max_x
max limit for x axis

extended_ticks
extend ticks on x axis?

xax_format
how to format tick labels. Currently one of "plain", "comma" or "date"

show_secondary_x_label
determines whether to show the year, or another unit of time in the case of smaller series, on the x-axis below the x-axis labels.

rug
show a "rug" plot next to the x axis? (default: FALSE - no)

Note

xax_format is likely to undergo a drastic change in future releases but support for these three formats will also likely remain.

[Package metricsgraphics version 0.9.1 Index]