mjs_grid {metricsgraphics}	R Documentation

Lays out metricsgraphics widgets into a "grid", similar to grid.arrange from gridExtra

Description

Lays out metricsgraphics widgets into a "grid", similar to grid.arrange from gridExtra

Usage

mjs_grid(..., ncol = 1, nrow = 1, widths = 1)

Arguments

...
either individual metricsgraphics objects or a mixture of individual metricsgraphics objects and lists of metricsgrahics objects.

ncol
how many columns in the grid

nrow
how many rows in the grid

widths
widths of the cells per row

Value

htmltools tag with wrapped metricsgraphics objects suitable for knitting with echo=FALSE & results="asis" or rendering in Viewer with html_print

Note

mjs_grid does not work in a Shiny context

[Package metricsgraphics version 0.9.1 Index]