mjs_hist {metricsgraphics}	R Documentation

Shortcut for plotting MetricsGraphics histograms

Description

This function performs the call to mjs_plot and assumes data is a numeric vector. It's intended to save keystrokes when plotting quick histograms. This function automatically a y axis label "Frequency" which you can override with a call to mjs_labs.

Usage

mjs_hist(data, bins = NULL, bar_margin = 1)

Arguments

data
numeric vector

bins
number of bins for the histogram (NULL == let MetricsGraphcis.js library compute)

bar_margin
space between bars (defaults to 1)

Value

metricsgraphics object

Examples

bimod <- c(rnorm(1000, 0, 1), rnorm(1000, 3, 1))

mjs_plot(bimod) %>% mjs_histogram()
bimod %>% mjs_hist()

mjs_plot(bimod) %>% mjs_histogram(bins=30)
bimod %>% mjs_hist(30)

[Package metricsgraphics version 0.9.1 Index]